requirements.txt
Cython==0.27.3
Pillow==4.3.0
PyWavelets==0.5.2
PyYAML==3.12
argparse==1.2.1
backports.functools-lru-cache==1.4
chardet==2.0.1
colorama==0.2.5
cycler==0.10.0
decorator==4.1.2
h5py==2.7.1
html5lib==0.999
leveldb==0.194
lmdb==0.93
matplotlib==2.1.0
networkx==2.0
nose==1.3.7
numpy==1.13.3
olefile==0.44
pandas==0.21.0
protobuf==3.4.0
pydot==1.2.3
pyparsing==2.2.0
python-dateutil==1.5
python-gflags==3.1.2
pytz==2017.3
requests==2.2.1
scikit-image==0.13.1
scipy==0.13.3
six==1.11.0
subprocess32==3.2.7
tqdm==4.19.4
urllib3==1.7.1
wsgiref==0.1.2
